+++ examples/eio/list_prefix.ml
@@ -19,8 +19,6 @@ let unwrap label = function
   | Ok value -> value
   | Error error -> fail "%s: %a" label Awskit_s3.Error.pp error
 
-let bucket_name value = Awskit_s3.Bucket_name.of_string_exn value
-
 let or_fail_msg label = function
   | Ok value -> value
   | Error (`Msg message) -> fail "%s: %s" label message
@@ -57,11 +55,8 @@ let create_s3 stdenv sw =
 
 let run stdenv =
   Eio.Switch.run @@ fun sw ->
-  let bucket = bucket_name (env "AWSKIT_EXAMPLE_BUCKET") in
-  let prefix =
-    Awskit_s3.Object_key.Prefix.of_string_exn
-      (env_default "AWSKIT_EXAMPLE_PREFIX" "awskit-examples/")
-  in
+  let bucket = env "AWSKIT_EXAMPLE_BUCKET" in
+  let prefix = env_default "AWSKIT_EXAMPLE_PREFIX" "awskit-examples/" in
   let s3 = create_s3 stdenv sw in
   let options = Awskit_s3.Object.List.options_exn ~prefix ~max_keys:25 () in
   let objects =