+++ examples/eio/put_get.ml
@@ -19,9 +19,6 @@ let unwrap label = function
   | Ok value -> value
   | Error error -> fail "%s: %a" label Awskit_s3.Error.pp error
 
-let bucket_name value = Awskit_s3.Bucket_name.of_string_exn value
-let object_key value = Awskit_s3.Object_key.of_string_exn value
-
 let or_fail_msg label = function
   | Ok value -> value
   | Error (`Msg message) -> fail "%s: %s" label message
@@ -58,10 +55,8 @@ let create_s3 stdenv sw =
 
 let run stdenv =
   Eio.Switch.run @@ fun sw ->
-  let bucket = bucket_name (env "AWSKIT_EXAMPLE_BUCKET") in
-  let key =
-    object_key (env_default "AWSKIT_EXAMPLE_KEY" "awskit-examples/put-get.txt")
-  in
+  let bucket = env "AWSKIT_EXAMPLE_BUCKET" in
+  let key = env_default "AWSKIT_EXAMPLE_KEY" "awskit-examples/put-get.txt" in
   let body = env_default "AWSKIT_EXAMPLE_BODY" "Hello from awskit live S3." in
   let s3 = create_s3 stdenv sw in
   let put =