+++ examples/lwt/list_prefix.ml
@@ -20,15 +20,11 @@ let unwrap label = function
   | Ok value -> value
   | Error error -> fail "%s: %a" label Awskit_s3.Error.pp error
 
-let bucket_name value = Awskit_s3.Bucket_name.of_string_exn value
 let create_s3 () = S3.create () |> unwrap "create S3 client"
 
 let run () =
-  let bucket = bucket_name (env "AWSKIT_EXAMPLE_BUCKET") in
-  let prefix =
-    Awskit_s3.Object_key.Prefix.of_string_exn
-      (env_default "AWSKIT_EXAMPLE_PREFIX" "awskit-examples/")
-  in
+  let bucket = env "AWSKIT_EXAMPLE_BUCKET" in
+  let prefix = env_default "AWSKIT_EXAMPLE_PREFIX" "awskit-examples/" in
   let s3 = create_s3 () in
   let options = Awskit_s3.Object.List.options_exn ~prefix ~max_keys:25 () in
   let* objects = S3.Object.List.objects s3 ~bucket ~options ~max_pages:4 () in