+++ examples/lwt/object_metadata.ml
@@ -20,8 +20,6 @@ let unwrap label = function
   | Ok value -> value
   | Error error -> fail "%s: %a" label Awskit_s3.Error.pp error
 
-let bucket_name value = Awskit_s3.Bucket_name.of_string_exn value
-let object_key value = Awskit_s3.Object_key.of_string_exn value
 let create_s3 () = S3.create () |> unwrap "create S3 client"
 
 let put_options =
@@ -39,10 +37,8 @@ let put_options =
     ()
 
 let run () =
-  let bucket = bucket_name (env "AWSKIT_EXAMPLE_BUCKET") in
-  let key =
-    object_key (env_default "AWSKIT_EXAMPLE_KEY" "awskit-examples/metadata.txt")
-  in
+  let bucket = env "AWSKIT_EXAMPLE_BUCKET" in
+  let key = env_default "AWSKIT_EXAMPLE_KEY" "awskit-examples/metadata.txt" in
   let s3 = create_s3 () in
   let* put =
     S3.Object.put_string s3 ~bucket ~key ~options:put_options