+++ examples/lwt/transfer_file.ml
@@ -20,8 +20,6 @@ let unwrap label = function
   | Ok value -> value
   | Error error -> fail "%s: %a" label Awskit_s3.Error.pp error
 
-let bucket_name value = Awskit_s3.Bucket_name.of_string_exn value
-let object_key value = Awskit_s3.Object_key.of_string_exn value
 let create_s3 () = S3.create () |> unwrap "create S3 client"
 
 let strategy_to_string = function
@@ -39,12 +37,11 @@ let progress label (event : Awskit_s3.Transfer.progress) =
       Format.printf "%s %Ld/%Ld bytes@." label event.transferred total
 
 let run () =
-  let bucket = bucket_name (env "AWSKIT_EXAMPLE_BUCKET") in
+  let bucket = env "AWSKIT_EXAMPLE_BUCKET" in
   let upload_path = env_default "AWSKIT_EXAMPLE_FILE" "README.md" in
   let key =
-    object_key
-      (env_default "AWSKIT_EXAMPLE_KEY"
-         ("awskit-examples/files/" ^ Filename.basename upload_path))
+    env_default "AWSKIT_EXAMPLE_KEY"
+      ("awskit-examples/files/" ^ Filename.basename upload_path)
   in
   let download_path =
     env_default "AWSKIT_EXAMPLE_DOWNLOAD"